+++ src/el/sayid.el
@@ -4,7 +4,7 @@
 
 ;; Author: Bill Piel <bill@billpiel.com>
 ;; Maintainer: Bozhidar Batsov <bozhidar@batsov.dev>
-;; Version: 0.7.1
+;; Version: 0.8.0
 ;; URL: https://github.com/clojure-emacs/sayid
 ;; Package-Requires: ((emacs "28") (cider "2.0"))
 ;; Keywords: clojure, cider, debugger
@@ -53,11 +53,11 @@ The injected dependencies are most likely nREPL middlewares."
   :type 'boolean)
 
 (defconst sayid-version
-  "0.7.1"
+  "0.8.0"
   "The current version of sayid.")
 
 (defconst sayid-injected-plugin-version
-  "0.7.1"
+  "0.8.0"
   "The version of the sayid Lein plugin to be automatically injected.")
 
 (defface sayid-int-face '((t :inherit default))