+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## master (unreleased)
+## [0.8.0] - 2026-07-17
 
 ### Changes